thin::test::edge_case_sizes
This cannot *fully* test e.g. writing above the 2 GiB mark on 32-bit targets.
This is a limitation of LLVM/rustc, as I understand it.
Then again, the test harness likely doesn't have a memory space that big.
We at least test writing to the byte just before the 2 GiB boundary.
We also test writing >2 GiB on 64-bit targets, naturally.